// src/parity/utils/types.ts
export interface MapData {
  autoplayBgm: boolean;
  autoplayBgs: boolean;
  battleback1Name: string;
  battleback2Name: string;
  bgm: AudioFile;
  bgs: AudioFile;
  disableDashing: boolean;
  displayName: string;
  encounterList: Encounter[];
  encounterStep: number;
  height: number;
  width: number;
  parallaxLoopX: boolean;
  parallaxLoopY: boolean;
  parallaxName: string;
  parallaxShow: boolean;
  parallaxSx: number;
  parallaxSy: number;
  scrollType: number;
  specifyBattleback: boolean;
  tilesetId: number;
  note: string;
  data: number[];
  events: (MapEvent | null)[];
}
// ...
export interface MapEvent {
  id: number;
  name: string;
  note: string;
  pages: EventPage[];
  x: number;
  y: number;
}

export interface EventPage {
  conditions: EventConditions;
  directionFix: boolean;
  image: EventImage;
  list: EventCommand[];
  moveFrequency: number;
  moveRoute: MoveRoute;
  moveSpeed: number;
  moveType: number;
  priorityType: number;
  stepAnime: boolean;
  through: boolean;
  trigger: number;
  walkAnime: boolean;
}
// ...
export interface EventCommand {
  code: number;
  indent: number;
  parameters: unknown[];
}

export interface MoveRoute {
  list: MoveCommand[];
  repeat: boolean;
  skippable: boolean;
  wait: boolean;
}

export interface MoveCommand {
  code: number;
  parameters: unknown[];
}

export interface AudioFile {
  name: string;
  pan: number;
  pitch: number;
  volume: number;
}

export interface Encounter {
  regionSet: number[];
  troopId: number;
  weight: number;
}